% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funkcja1.R
\name{Funkcja1}
\alias{Funkcja1}
\title{Obliczanie emisji}
\usage{
Funkcja1(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = "Diesel",
  euro = "Euro 4",
  technologia = "DPF",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Obliczanie emisji
}
