% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funkcja1.R
\name{Funkcja1}
\alias{Funkcja1}
\title{Funkcja oblicza emisje dla aut osobowych poruszajacych sie na diesel, z technologia DPF.
Pod uwage brane sa zanieczyszczenia EC oraz CO.}
\usage{
Funkcja1(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = "Diesel",
  euro = "Euro 4",
  technologia = "DPF",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{dane}{dane wejsciowe}

\item{kategoria}{rodzaj auta}

\item{paliwo}{rodzaj paliwa}

\item{euro}{euro}

\item{mode}{mode}

\item{substancja}{zanieczyszczenie}
}
\value{

}
\description{
Funkcja oblicza emisje dla aut osobowych poruszajacych sie na diesel, z technologia DPF.
Pod uwage brane sa zanieczyszczenia EC oraz CO.
}
