% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funkcja2.R
\name{Funkcja2}
\alias{Funkcja2}
\title{Wykres
Przedstawia on zależnosc emisji od natezenia ruchu}
\usage{
Funkcja2(
  dane = input,
  kategoria = c("Passenger Cars", "Light Commercial Vehicles", "Heavy Duty Trucks",
    "Buses"),
  paliwo = c("Petrol", "Diesel", "Biodiesel", "Petrol Hybrid", "CNG"),
  euro = c("Euro 5", "Euro 4"),
  technologia = c("PFI", "GDI", "DPF + SCR", "EGR", "DPF"),
  mode = "",
  substancja = c("EC", "CO", "NOx", "CH4")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Wykres
Przedstawia on zależnosc emisji od natezenia ruchu
}
\examples{
Funkcja2()
}
