% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funkcja2.R
\name{Funkcja2}
\alias{Funkcja2}
\title{Title}
\usage{
Funkcja2(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = "Petrol",
  euro = "Euro 5",
  technologia = "GDI",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Title
}
