% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funkcja2.R
\name{Funkcja2}
\alias{Funkcja2}
\title{Wykres
Przedstawia on zależnosc emisji od natezenia ruchu}
\usage{
Funkcja2(
  dane = input,
  kategoria = "Passenger Cars",
  paliwo = "Diesel",
  euro = "Euro 4",
  technologia = "DPF",
  mode = "",
  substancja = c("EC", "CO")
)
}
\arguments{
\item{substancja}{}
}
\value{

}
\description{
Wykres
Przedstawia on zależnosc emisji od natezenia ruchu
}
\examples{
Funkcja2()
}
